feat(server): add `server` property to page meta
Adds the `server` feature for the page head so we can perform operations on the server side only. This is useful when we want to load content from sources and set their results as variables in the context so the page loads from the server with the content populated. This increases the page load speed and is helpful to perform operations that are server-side only.