fix: templates keep placeholders from app.render
The placeholders are necessary to be kept so we can properly parse and compile them at run time for the "control" components like loop. Loop components can receive parsed variables that are not solved yet and are not in the context yet, they will be present only at run time.